requirements.txt
streamlit
Pillow
base64
numba
numpy
scipy
pandas
py3Dmol
stmol==0.0.7
st-annotated-text
scikit-learn
networkx
plotly
matplotlib
maturin
flask
flask-cors
gunicorn
bio
requests